requirements.txt
streamlit==1.29.0
cohere==4.37.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
python-dotenv==1.0.0
plotly==5.17.0
chromadb==0.4.15
sentence-transformers==2.2.2
huggingface-hub==0.16.4
langchain==0.1.0
langchain-community==0.0.10
